// src/app/shared/enums/role.ts
export enum Role {
    Reseller = 'Reseller',
    Marketing = 'Marketing',
    Carrier = 'Customer',
    CustomerSVC = 'Customer Service',
    CustomerSvcMgr = 'Customer Service Manager',
    CustomerServiceExt = 'Customer Service (External)',
    SuperAdmin = 'Super Admin',
    Finance = 'Finance',
    BIC = 'BIC',
    RailInc = 'RailInc',
    Border = 'Border',
    ExecDirector = 'Executive Director',
    UIIA = 'UIIA'
}